+++ src/client.ts
@@ -1,121 +1,236 @@
-import http from 'http';
-
-import { IntegrationProviderAuthenticationError } from '@jupiterone/integration-sdk-core';
+import fetch, { Response } from 'node-fetch';
+import {
+  IntegrationProviderAPIError,
+  IntegrationProviderAuthenticationError,
+} from '@jupiterone/integration-sdk-core';
+import { retry } from '@lifeomic/attempt';
 
 import { IntegrationConfig } from './config';
-import { AcmeUser, AcmeGroup } from './types';
+import { JenkinsUser, JenkinsRole, JenkinsBuild, JenkinsJob } from './types';
 
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
-/**
- * An APIClient maintains authentication state and provides an interface to
- * third party data APIs.
- *
- * It is recommended that integrations wrap provider data APIs to provide a
- * place to handle error responses and implement common patterns for iterating
- * resources.
- */
 export class APIClient {
   constructor(readonly config: IntegrationConfig) {}
 
-  public async verifyAuthentication(): Promise<void> {
-    // TODO make the most light-weight request possible to validate
-    // authentication works with the provided credentials, throw an err if
-    // authentication fails
-    const request = new Promise<void>((resolve, reject) => {
-      http.get(
-        {
-          hostname: 'localhost',
-          port: 443,
-          path: '/api/v1/some/endpoint?limit=1',
-          agent: false,
-          timeout: 10,
+  private baseUri = this.config.hostName;
+  private withBaseUri = (path: string) => `${this.baseUri}${path}`;
+  private roleTypes = ['globalRoles', 'projectRoles', 'slaveRoles'];
+  private perPage = 50;
+
+  private checkStatus = (response: Response) => {
+    if (response.ok) {
+      return response;
+    } else {
+      throw new IntegrationProviderAPIError(response);
+    }
+  };
+
+  private async getRequest(endpoint: string, method: 'GET'): Promise<Response> {
+    const auth =
+      'Basic ' +
+      Buffer.from(
+        `${this.config.userName}` + ':' + `${this.config.apiKey}`,
+      ).toString('base64');
+    try {
+      const options = {
+        method,
+        headers: {
+          Authorization: auth,
         },
-        (res) => {
-          if (res.statusCode !== 200) {
-            reject(new Error('Provider authentication failed'));
-          } else {
-            resolve();
-          }
+      };
+
+      const response = await retry(
+        async () => {
+          const res: Response = await fetch(endpoint, options);
+          this.checkStatus(res);
+          return res;
+        },
+        {
+          delay: 5000,
+          factor: 2,
+          maxAttempts: 5,
+          minDelay: 100,
+          maxDelay: 500,
+          jitter: true,
+          handleError: (err, context) => {
+            if (
+              err.statusCode !== 429 ||
+              ([500, 502, 503].includes(err.statusCode) &&
+                context.attemptNum > 1)
+            ) {
+              context.abort();
+            }
+          },
         },
       );
-    });
 
-    try {
-      await request;
+      return response.json();
     } catch (err) {
-      throw new IntegrationProviderAuthenticationError({
-        cause: err,
-        endpoint: 'https://localhost/api/v1/some/endpoint?limit=1',
+      throw new IntegrationProviderAPIError({
+        endpoint: endpoint,
         status: err.status,
         statusText: err.statusText,
       });
     }
   }
 
-  /**
-   * Iterates each user resource in the provider.
-   *
-   * @param iteratee receives each resource to produce entities/relationships
-   */
-  public async iterateUsers(
-    iteratee: ResourceIteratee<AcmeUser>,
+  private async paginatedGeneralRequest<T>(
+    uri: string,
+    method: 'GET',
+    accessor: string,
+    iteratee: ResourceIteratee<T>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
+    let numElements = 0;
+    let start = 0;
+    let end = this.perPage;
 
-    const users: AcmeUser[] = [
-      {
-        id: 'acme-user-1',
-        name: 'User One',
-      },
-      {
-        id: 'acme-user-2',
-        name: 'User Two',
-      },
-    ];
+    do {
+      const response = await this.getRequest(`${uri}{${start},${end}}`, method);
+      for (const item of response[accessor]) {
+        await iteratee(item);
+      }
 
-    for (const user of users) {
-      await iteratee(user);
+      numElements = response[accessor].length;
+      start = end - 1;
+      end = end + this.perPage;
+    } while (numElements == this.perPage);
+  }
+
+  private async paginatedBuildRequest<T>(
+    uri: string,
+    method: 'GET',
+    iteratee: ResourceIteratee<T>,
+  ): Promise<void> {
+    let buildLength = 0;
+    let start = 0;
+    let end = this.perPage;
+    do {
+      const response = await this.getRequest(
+        `${uri}tree=allBuilds[url]{${start},${end}}`,
+        method,
+      );
+
+      for (const item of response.allBuilds || []) {
+        await iteratee(item);
+      }
+
+      if (response.allBuilds) {
+        buildLength = response.allBuilds.length;
+      }
+
+      start = end;
+      end += this.perPage;
+    } while (buildLength);
+  }
+
+  public async verifyAuthentication(): Promise<void> {
+    const uri = this.withBaseUri(`/api/json`);
+    try {
+      await this.getRequest(uri, 'GET');
+    } catch (err) {
+      throw new IntegrationProviderAuthenticationError({
+        cause: err,
+        endpoint: uri,
+        status: err.status,
+        statusText: err.statusText,
+      });
     }
   }
 
-  /**
-   * Iterates each group resource in the provider.
-   *
-   * @param iteratee receives each resource to produce entities/relationships
-   */
-  public async iterateGroups(
-    iteratee: ResourceIteratee<AcmeGroup>,
+  public async iterateUsers(
+    iteratee: ResourceIteratee<JenkinsUser>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
+    await this.paginatedGeneralRequest<JenkinsUser>(
+      this.withBaseUri(
+        `/asynchPeople/api/json?tree=users[lastChange,project[name,url],user[absoluteUrl,fullName]]`,
+      ),
+      'GET',
+      'users',
+      iteratee,
+    );
+  }
 
-    const groups: AcmeGroup[] = [
-      {
-        id: 'acme-group-1',
-        name: 'Group One',
-        users: [
-          {
-            id: 'acme-user-1',
-          },
-        ],
-      },
-    ];
+  public async iterateJobs(
+    iteratee: ResourceIteratee<JenkinsJob>,
+  ): Promise<void> {
+    await this.paginatedGeneralRequest<JenkinsJob>(
+      this.withBaseUri(`/api/json?tree=jobs[_class,url,name,color]`),
+      'GET',
+      'jobs',
+      iteratee,
+    );
+  }
 
-    for (const group of groups) {
-      await iteratee(group);
+  public async iterateSubJobs(
+    iteratee: ResourceIteratee<JenkinsJob>,
+    url: string,
+  ): Promise<void> {
+    await this.paginatedGeneralRequest<JenkinsJob>(
+      `${url}api/json?tree=jobs[_class,url,name,color]`,
+      'GET',
+      'jobs',
+      iteratee,
+    );
+  }
+
+  public async iterateBuilds(
+    iteratee: ResourceIteratee<JenkinsBuild>,
+    projectUrl: string,
+  ): Promise<void> {
+    await this.paginatedBuildRequest<JenkinsBuild>(
+      `${projectUrl}api/json?`,
+      'GET',
+      iteratee,
+    );
+  }
+
+  public async fetchBuildDetails(url: string): Promise<JenkinsBuild> {
+    return await this.getRequest(`${url}/api/json`, 'GET');
+  }
+
+  public async listAllRoles(): Promise<any> {
+    const roleList: Array<string> = [];
+
+    for (const roleType of this.roleTypes) {
+      const roles = await this.getRequest(
+        this.withBaseUri(
+          `/role-strategy/strategy/getAllRoles?type=${roleType}`,
+        ),
+        'GET',
+      );
+      for (const role in roles) {
+        roleList.push(`${roleType}:${role}`);
+      }
     }
+
+    return roleList;
+  }
+
+  public async getRoleMembers(role: JenkinsRole): Promise<any> {
+    const request = await this.getRequest(
+      this.withBaseUri(
+        `/role-strategy/strategy/getRole?type=${role.roleType}&roleName=${role.name}`,
+      ),
+      'GET',
+    );
+    return request.sids;
+  }
+
+  public async getRoleDetails(role: string): Promise<JenkinsRole> {
+    const colonIndex = role.lastIndexOf(':');
+    const roleName = role.substring(colonIndex + 1);
+    const roleType = role.substring(0, colonIndex);
+
+    const request = await this.getRequest(
+      this.withBaseUri(
+        `/role-strategy/strategy/getRole?type=${roleType}&roleName=${roleName}`,
+      ),
+      'GET',
+    );
+    request.name = roleName;
+    request.roleType = `${roleType}`;
+    return request;
   }
 }