+++ test/config.ts
@@ -2,6 +2,7 @@ import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 import { StepTestConfig } from '@jupiterone/integration-sdk-testing';
 import * as dotenv from 'dotenv';
 import * as path from 'path';
+
 import { invocationConfig } from '../src';
 import { IntegrationConfig } from '../src/config';
 
@@ -10,12 +11,14 @@ if (process.env.LOAD_ENV) {
     path: path.join(__dirname, '../.env'),
   });
 }
-const DEFAULT_CLIENT_ID = 'dummy-acme-client-id';
-const DEFAULT_CLIENT_SECRET = 'dummy-acme-client-secret';
+const DEFAULT_USERNAME = 'dummy_username';
+const DEFAULT_CLIENT_SECRET = 'dummy_secret';
+const DEFAULT_HOSTNAME = 'http://localhost:8080';
 
 export const integrationConfig: IntegrationConfig = {
-  clientId: process.env.CLIENT_ID || DEFAULT_CLIENT_ID,
-  clientSecret: process.env.CLIENT_SECRET || DEFAULT_CLIENT_SECRET,
+  userName: process.env.USER_NAME || DEFAULT_USERNAME,
+  apiKey: process.env.API_KEY || DEFAULT_CLIENT_SECRET,
+  hostName: process.env.HOST_NAME || DEFAULT_HOSTNAME,
 };
 
 export function buildStepTestConfigForStep(stepId: string): StepTestConfig {